add redirect modal to Global guest modal, trigger modal
