+++ packages/ui/src/validators/components/ValidatorItem.tsx
@@ -8,8 +8,9 @@ import { TableListItemAsLinkHover } from '@/common/components/List'
 import { Skeleton } from '@/common/components/Skeleton'
 import { TextMedium, TokenValue } from '@/common/components/typography'
 import { BorderRad, Colors, Sizes, Transitions } from '@/common/constants'
-import { useLocalStorage } from '@/common/hooks/useLocalStorage'
+import { useModal } from '@/common/hooks/useModal'
 
+import { NominatingRedirectModalCall } from '../modals/NominatingRedirectModal'
 import { ValidatorWithDetails } from '../types/Validator'
 
 import { ValidatorInfo } from './ValidatorInfo'
@@ -20,10 +21,8 @@ interface ValidatorItemProps {
 }
 export const ValidatorItem = ({ validator, onClick }: ValidatorItemProps) => {
   const { stashAccount, membership, isVerifiedValidator, isActive, commission, APR, staking } = validator
-  const [, setShow] = useLocalStorage<boolean>('NominateRedirectModal')
-  const showNominateRedirectModal = () => {
-    setShow(true)
-  }
+  const { showModal } = useModal<NominatingRedirectModalCall>()
+
   return (
     <ValidatorItemWrapper onClick={onClick}>
       <ValidatorItemWrap>
@@ -42,7 +41,13 @@ export const ValidatorItem = ({ validator, onClick }: ValidatorItemProps) => {
         <TokenValue size="xs" value={staking.total} />
         <TextMedium bold>{APR}%</TextMedium>
         <TextMedium bold>{commission}%</TextMedium>
-        <ButtonPrimary size="small" onClick={showNominateRedirectModal}>
+        <ButtonPrimary
+          size="small"
+          onClick={(e) => {
+            e.stopPropagation()
+            showModal({ modal: 'NominatingRedirect' })
+          }}
+        >
           Nominate
         </ButtonPrimary>
       </ValidatorItemWrap>
