+++ packages/ui/src/validators/modals/NominatingRedirectModal/index.ts
@@ -1,0 +1,2 @@
+export * from './NominatingRedirectModal'
+export * from './types'
