+++ packages/ui/src/validators/modals/NominatingRedirectModal/types.ts
@@ -1,0 +1,3 @@
+import { ModalCall } from '../../../common/providers/modal/types'
+
+export type NominatingRedirectModalCall = ModalCall<'NominatingRedirect'>
